Move set up steps guard into a separate folder.
