+++ console/src/app/modules/mongoose-set-up-module/set-up-routing/mongoose-setup.routes.ts
@@ -4,7 +4,7 @@ import { MongooseSetUpComponent } from "../components/mongoose-set-up/mongoose-s
 import { NodesComponent } from "../components/mongoose-set-up/set-up-steps/nodes/nodes.component";
 import { ConfigurationEditingRootComponent } from "../components/mongoose-set-up/set-up-steps/configuration-set-up/control-editing-root/control-editing-root.component";
 import { ScenariosComponent } from "../components/mongoose-set-up/set-up-steps/scenarios-set-up/scenarios/scenarios.component";
-import { SetupStepsGuard } from "src/app/setup-steps.guard";
+import { SetupStepsGuard } from "src/app/modules/mongoose-set-up-module/set-up-routing/guards/setup-steps.guard";
 
 export const SETUP_ROUTES: Routes = [
   { path: RoutesList.NODES, component: NodesComponent },
